refactor: rewrite data access layer to use repository pattern

+++ part3/app/api/v1/reviews.py
@@ -1,8 +1,8 @@
+from flask_jwt_extended import get_jwt_identity, jwt_required
 from flask_restx import Namespace, Resource, fields
+from werkzeug.exceptions import Unauthorized
 
 from app.services import facade
-from app.models.review import validate_rating
-from flask_jwt_extended import jwt_required, get_jwt_identity
 
 api = Namespace('reviews', description='Review operations')
 
@@ -31,17 +31,22 @@ class ReviewList(Resource):
         # if not validate_rating(rating):
         #     return {'error': 'Invalid input data'}, 400
 
-        review = facade.create_review(review_data)
-        # if not review:
-        #     return {'error': 'Invalid input data'}, 400
+        try:
+            from app.schemas.review_schema import ReviewCreateSchema
+            ReviewCreateSchema.validate(review_data)
 
-        return {
-            'id': review.id,
-            'comment': review.comment,
-            'rating': review.rating,
-            'user_id': review.user.id,
-            'place_id': review.place.id
-        }, 201
+            review = facade.create_review(review_data)
+
+            return {
+                'id': review.id,
+                'comment': review.comment,
+                'rating': review.rating,
+                'user_id': review.user.id,
+                'place_id': review.place.id
+            }, 201
+        except ValueError as e:
+            return {'message': str(e)}, 400
+
 
     @api.response(200, 'List of reviews retrieved successfully')
     def get(self):
@@ -75,24 +80,48 @@ class ReviewResource(Resource):
 
     @api.expect(review_model)
     @api.response(200, 'Review updated successfully')
-    @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @api.response(403, 'Unauthorized action')
+    @api.response(404, 'Review not found')
     def put(self, review_id):
         """Update a review's information"""
+        current_user_id = get_jwt_identity()
+        current_user = get_jwt()
+
         review_data = api.payload
-        review = facade.update_review(review_id, review_data)
-        return {
-            'id': review.id,
-            'comment': review.comment,
-            'rating': review.rating
-        }, 200
+
+        if (not current_user.get('is_admin')or
+                not str(current_user_id).strip() ==
+                str(review_data['user_id']).strip()):
+            raise Unauthorized('Unauthorized action')
+        try:
+            from app.schemas.review_schema import ReviewUpdateSchema
+            ReviewUpdateSchema.validate(review_data)
+
+            review = facade.update_review(review_id, review_data)
+
+            return {
+                'id': review.id,
+                'comment': review.comment,
+                'rating': review.rating
+            }, 200
+        except ValueError as e:
+            return {'message': str(e)}, 400
+
 
     @api.response(200, 'Review deleted successfully')
+    @api.response(403, 'Unauthorized action')
     @api.response(404, 'Review not found')
     def delete(self, review_id):
         """Delete a review"""
+        current_user_id = get_jwt_identity()
+
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
+
+        if not str(current_user_id).strip() == str(review['user_id']).strip():
+            raise Unauthorized('Unauthorized action')
+
         facade.delete_review(review_id)
         return {'message': 'Review deleted successfully'}, 200
